Change StatsService.java

+++ src/main/java/ru/netology/stats/StatsService.java
@@ -3,12 +3,13 @@ package ru.netology.stats;
 public class StatsService {
 
     // Сумма продаж
+
     public long sumSales(long[] sales) {
         long allSales = 0;
         for (long s : sales) {
             allSales += s;
         }
-        return  allSales;
+        return allSales;
     }
 /*      long allSales = 0;
         for (int i = 0; i < sales.length; i++) {
@@ -19,6 +20,7 @@ public class StatsService {
 */
 
     // Средняя сумма продаж в месяц
+
     public long averageSales(long[] sales) {
         long averageValue = 0;
         for (long s : sales) {
@@ -32,7 +34,7 @@ public class StatsService {
             averageValue +=  sales[i];
         }
         averageValue = averageValue / sales.length;
-        return (int) averageValue;
+        return  averageValue;
     }
 */
 
@@ -41,7 +43,7 @@ public class StatsService {
     public int maxSales(long[] sales) {
         int maxMonth = 0;
         for (int i = 0; i < sales.length; i++) {
-            if (sales[i] >=  sales[maxMonth]) {
+            if (sales[i] >= sales[maxMonth]) {
                 maxMonth = i;
             }
         }
@@ -49,10 +51,11 @@ public class StatsService {
     }
 
     //  Месяц с минимальными продажами
+
     public int minSales(long[] sales) {
         int minMonth = 0;
         for (int i = 0; i < sales.length; i++) {
-            if (sales[i] <= (int) sales[minMonth]) {
+            if (sales[i] <= sales[minMonth]) {
                 minMonth = i;
             }
         }
@@ -63,7 +66,7 @@ public class StatsService {
     public int belowAverageMonths(long[] sales) {
 
         int count = 0;
-        long middleValue =  averageSales(sales);
+        long middleValue = averageSales(sales);
 
         for (int i = 0; i < sales.length; i++) {
             if (sales[i] < middleValue) {
@@ -77,7 +80,7 @@ public class StatsService {
     public int aboveAverageMonths(long[] sales) {
 
         int count = 0;
-        long middleValue =  averageSales(sales);
+        long middleValue = averageSales(sales);
 
         for (int i = 0; i < sales.length; i++) {
             if (sales[i] > middleValue) {
